requirements.txt
beautifulsoup4==4.13.4
duckduckgo_search==8.0.2
langchain==0.3.25
langchain_core==0.3.62
langchain_groq==0.3.2
python-dotenv==1.1.0
Requests==2.32.3
streamlit==1.45.1
